+++ server/src/main/java/org/jgrasstools/server/jetty/fileupload/FileUploadProgressListener.java
@@ -1,0 +1,91 @@
+package org.jgrasstools.server.jetty.fileupload;
+import org.apache.commons.fileupload.ProgressListener;
+
+public class FileUploadProgressListener implements ProgressListener {
+
+    private long num100Ks = 0;
+
+    private long theBytesRead = 0;
+    private long theContentLength = -1;
+    private int whichItem = 0;
+    private int percentDone = 0;
+    private boolean contentLengthKnown = false;
+
+    public void update( long bytesRead, long contentLength, int items ) {
+
+        if (contentLength > -1) {
+            contentLengthKnown = true;
+        }
+        theBytesRead = bytesRead;
+        theContentLength = contentLength;
+        whichItem = items;
+
+        long nowNum100Ks = bytesRead / 100000;
+        // Only run this code once every 100K
+        if (nowNum100Ks > num100Ks) {
+            num100Ks = nowNum100Ks;
+            if (contentLengthKnown) {
+                percentDone = (int) Math.round(100.00 * bytesRead / contentLength);
+            }
+            System.out.println(getMessage());
+        }
+    }
+
+    public String getMessage() {
+        if (theContentLength == -1) {
+            return "" + theBytesRead + " of Unknown-Total bytes have been read.";
+        } else {
+            return "" + theBytesRead + " of " + theContentLength + " bytes have been read (" + percentDone + "% done).";
+        }
+
+    }
+
+    public long getNum100Ks() {
+        return num100Ks;
+    }
+
+    public void setNum100Ks( long num100Ks ) {
+        this.num100Ks = num100Ks;
+    }
+
+    public long getTheBytesRead() {
+        return theBytesRead;
+    }
+
+    public void setTheBytesRead( long theBytesRead ) {
+        this.theBytesRead = theBytesRead;
+    }
+
+    public long getTheContentLength() {
+        return theContentLength;
+    }
+
+    public void setTheContentLength( long theContentLength ) {
+        this.theContentLength = theContentLength;
+    }
+
+    public int getWhichItem() {
+        return whichItem;
+    }
+
+    public void setWhichItem( int whichItem ) {
+        this.whichItem = whichItem;
+    }
+
+    public int getPercentDone() {
+        return percentDone;
+    }
+
+    public void setPercentDone( int percentDone ) {
+        this.percentDone = percentDone;
+    }
+
+    public boolean isContentLengthKnown() {
+        return contentLengthKnown;
+    }
+
+    public void setContentLengthKnown( boolean contentLengthKnown ) {
+        this.contentLengthKnown = contentLengthKnown;
+    }
+
+}
