+++ server/src/main/java/org/jgrasstools/server/jetty/fileupload/ProgressServlet.java
@@ -9,29 +9,30 @@ import javax.servlet.http.HttpSession;
 
 public class ProgressServlet extends HttpServlet {
 
-	private static final long serialVersionUID = 1L;
+    private static final long serialVersionUID = 1L;
 
-	public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-		doPost(request, response);
-	}
+    public void doGet( HttpServletRequest request, HttpServletResponse response ) throws IOException {
+        doPost(request, response);
+    }
 
-	public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
-		response.setContentType("text/html");
-		PrintWriter out = response.getWriter();
+    public void doPost( HttpServletRequest request, HttpServletResponse response ) throws IOException {
+        response.setContentType("text/html");
+        PrintWriter out = response.getWriter();
 
-		HttpSession session = request.getSession(true);
-		if (session == null) {
-			out.println("Sorry, session is null"); // just to be safe
-			return;
-		}
+        HttpSession session = request.getSession(true);
+        if (session == null) {
+            out.println("Sorry, session is null"); // just to be safe
+            return;
+        }
 
-		TestProgressListener testProgressListener = (TestProgressListener) session.getAttribute("testProgressListener");
-		if (testProgressListener == null) {
-			out.println("Progress listener is null");
-			return;
-		}
+        FileUploadProgressListener fileUploadProgressListener = (FileUploadProgressListener) session.getAttribute("testProgressListener");
+        if (fileUploadProgressListener == null) {
+            out.println("Progress listener is null");
+            return;
+        }
 
-		out.println(testProgressListener.getMessage());
+        // out.println(testProgressListener.getMessage());
+        out.println(fileUploadProgressListener.getPercentDone());
 
-	}
+    }
 }
