+++ server/src/main/webapp/WEB-INF/web.xml
@@ -4,16 +4,16 @@
 	xsi:schemaLocation="http://java.sun.com/xml/ns/j2ee http://java.sun.com/xml/ns/j2ee/web-app_2_4.xsd">
 
 	<servlet>
-		<servlet-name>TestServlet</servlet-name>
-		<servlet-class>org.jgrasstools.server.jetty.example1.TestServlet</servlet-class>
+		<servlet-name>FileUpoadServlet</servlet-name>
+		<servlet-class>org.jgrasstools.server.jetty.fileupload.FileUpoadServlet</servlet-class>
 	</servlet>
 	<servlet>
 		<servlet-name>ProgressServlet</servlet-name>
-		<servlet-class>org.jgrasstools.server.jetty.example1.ProgressServlet</servlet-class>
+		<servlet-class>org.jgrasstools.server.jetty.fileupload.ProgressServlet</servlet-class>
 	</servlet>
 	<servlet-mapping>
-		<servlet-name>TestServlet</servlet-name>
-		<url-pattern>/test</url-pattern>
+		<servlet-name>FileUpoadServlet</servlet-name>
+		<url-pattern>/upload_file_progress</url-pattern>
 	</servlet-mapping>
 	<servlet-mapping>
 		<servlet-name>ProgressServlet</servlet-name>
