--- server/src/main/webapp/upload.jsp
@@ -1,18 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=ISO-8859-1" pageEncoding="ISO-8859-1"%>
-<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
-<html>
-<head>
-<meta http-equiv="Content-Type" content="text/html; charset=ISO-8859-1">
-<title>Upload Page</title>
-</head>
-<body>
-	<form name="form1" id="form1" action="file_upload" method="post" enctype="multipart/form-data">
-	<input type="hidden" name="hiddenfield1" value="ok">
-	Files to upload:
-	<br/>
-	<input type="file" size="50" name="file">
-	<br/>
-	<input type="submit" value="Upload">
-	</form>
-</body>
-</html>
